# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "The shape of an error returned by the API (FastAPI's default `detail` key).",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PostDetail": {
        "properties": {
          "body": {
            "description": "The post as Markdown.",
            "title": "Body",
            "type": "string"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "dek": {
            "description": "One-sentence summary shown under the title.",
            "title": "Dek",
            "type": "string"
          },
          "draft": {
            "description": "True only for a draft shown because APP_SHOW_DRAFTS is on.",
            "title": "Draft",
            "type": "boolean"
          },
          "read_minutes": {
            "description": "Estimated reading time in minutes.",
            "minimum": 1.0,
            "title": "Read Minutes",
            "type": "integer"
          },
          "slug": {
            "description": "The URL part: /writing/<slug>.",
            "title": "Slug",
            "type": "string"
          },
          "tag": {
            "title": "Tag",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "title",
          "dek",
          "tag",
          "date",
          "read_minutes",
          "draft",
          "body"
        ],
        "title": "PostDetail",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Personal Site API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/posts/{slug}": {
      "get": {
        "operationId": "get_post_api_posts__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "There is no post with this slug."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Post",
        "tags": [
          "posts"
        ]
      }
    },
...
  }
}